add danmu module and update.

+++ public/javascripts/test_adminManager.js
@@ -247,34 +247,43 @@ ManagerAction.updateTable = function (JSONdata) {
                 $frequencyTd.text(test.frequency);
 
                 /* 删除 */
-                var $deleteTd = $('<td></td>');
-                var $deleteSpan = $('<span class="glyphicon glyphicon-trash delete"></span>');
+                var $deleteTd = $('<td class="delete-td"></td>');
+                var $deleteSpan = $('<span class="glyphicon glyphicon-trash"></span>');
                 /* 删除事件绑定 */
                 $deleteSpan.click(function () {
+
                     var that = this;
-                    $.post('/test/deleteOne', {
-                        testTitle: test.testTitle,
-                        testType: test.testType
-                    }, function (JSONdata) {
-                        var JSONobject = JSON.parse(JSONdata);
-                        if(JSONobject.error){
-                            return ManagerAction.modalWindow("删除错误,请重试!");
-                        }
-                        $(that).parent().parent().remove();
-                    }, "JSON");
+                    if(confirm('确定删除: [ ' + test.testTitle + ' ]?')){
+
+                        $.post('/test/deleteOne', {
+                            testTitle: test.testTitle,
+                            testType: test.testType
+
+                        }, function (JSONdata) {
+                            var JSONobject = JSON.parse(JSONdata);
+                            if(JSONobject.error){
+                                return ManagerAction.modalWindow("删除错误,请重试!");
+                            }
+                            $(that).parent().parent().remove();
+
+                        }, "JSON");
+                    }
+
                 }).appendTo($deleteTd);
 
                 /* 编辑 */
-                var $editTd = $('<td></td>');
-                var $editSpan = $('<span class="glyphicon glyphicon-edit edit"></span>');
-                /* 编辑事件绑定 */
-                $editSpan.click(function () {
-                    var that = this;
-                }).appendTo($editTd);
+                // var $editTd = $('<td class="edit-td"></td>');
+                // var $editSpan = $('<span class="glyphicon glyphicon-edit"></span>');
+                // /* 编辑事件绑定 */
+                // $editSpan.click(function () {
+                //
+                //     window.open('/test/admin/edit/' + test.testType + '/' + test.testTitle);
+                //
+                // }).appendTo($editTd);
 
                 /* 预览 */
-                var $previewTd = $('<td>');
-                var $previewSpan = $('<span class="glyphicon glyphicon-eye-open preview">');
+                var $previewTd = $('<td class="preview-td">');
+                var $previewSpan = $('<span class="glyphicon glyphicon-eye-open">');
                 /* 预览事件绑定 */
                 $previewSpan.click(function () {
                     // 触发预览事件
@@ -285,7 +294,7 @@ ManagerAction.updateTable = function (JSONdata) {
                 $tr.append($titleTd).append($typeTd)
                     .append($dateTd).append($scoreModeTd)
                     .append($frequencyTd).append($previewTd)
-                    .append($editTd).append($deleteTd);
+                    .append($deleteTd);
 
                 // 表格添加
                 $testTable.append($tr);
